+++ spdivik/kmeans/__init__.py
@@ -1,5 +1,6 @@
 """Numpy-based implementation of k-means algorithm"""
 
+from ._auto import AutoKMeans
 from ._initialization import \
     ExtremeInitialization, \
     Initialization, \
